chore: 유지보수를 위해 외부 infra 도메인을 infra 하위 패키지로 이동
