--- src/main/java/com/groom/orbit/ai/app/dto/UpdateVectorGoalDto.java
@@ -1,6 +1,0 @@
-package com.groom.orbit.ai.app.dto;
-
-import lombok.Builder;
-
-@Builder
-public record UpdateVectorGoalDto(Long memberId, String goal, String newGoal) {}
