--- src/main/java/com/groom/orbit/ai/app/dto/UpdateVectorQuestDto.java
@@ -1,6 +1,0 @@
-package com.groom.orbit.ai.app.dto;
-
-import lombok.Builder;
-
-@Builder
-public record UpdateVectorQuestDto(Long memberId, String quest, String newQuest) {}
