--- src/main/java/com/groom/orbit/ai/dao/vector/Vector.java
@@ -1,13 +1,0 @@
-package com.groom.orbit.ai.dao.vector;
-
-import java.util.List;
-
-import lombok.Builder;
-
-@Builder
-public record Vector(
-    Long memberId,
-    String memberName,
-    List<String> interestJobs,
-    List<String> goals,
-    List<String> quests) {}
