--- src/main/java/com/groom/orbit/fcm/FcmMessageDto.java
@@ -1,8 +1,0 @@
-package com.groom.orbit.fcm;
-
-public record FcmMessageDto(Boolean validateOnly, Message message) {
-
-  public static FcmMessageDto of(Boolean validateOnly, Message message) {
-    return new FcmMessageDto(validateOnly, message);
-  }
-}
