--- src/main/java/com/groom/orbit/fcm/FcmSendRequestDto.java
@@ -1,8 +1,0 @@
-package com.groom.orbit.fcm;
-
-public record FcmSendRequestDto(String fcmToken, String title, String body) {
-
-  public static FcmSendRequestDto of(String fcmToken, String title, String body) {
-    return new FcmSendRequestDto(fcmToken, title, body);
-  }
-}
