--- src/main/java/com/groom/orbit/fcm/Message.java
@@ -1,8 +1,0 @@
-package com.groom.orbit.fcm;
-
-public record Message(Notification notification, String token) {
-
-  public static Message of(Notification notification, String token) {
-    return new Message(notification, token);
-  }
-}
