--- src/main/java/com/groom/orbit/fcm/SaveFcmTokenRequestDto.java
@@ -1,3 +1,0 @@
-package com.groom.orbit.fcm;
-
-public record SaveFcmTokenRequestDto(String fcmToken) {}
