+++ src/main/java/com/groom/orbit/goal/goal/app/MemberGoalService.java
@@ -10,9 +10,6 @@ import org.springframework.data.domain.Sort.Order;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.groom.orbit.ai.app.VectorService;
-import com.groom.orbit.ai.app.dto.CreateVectorDto;
-import com.groom.orbit.ai.app.dto.UpdateVectorGoalDto;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.common.exception.CommonException;
 import com.groom.orbit.common.exception.ErrorCode;
@@ -28,6 +25,9 @@ import com.groom.orbit.goal.goal.dao.entity.MemberGoal;
 import com.groom.orbit.goal.quest.app.dto.response.GetQuestResponseDto;
 import com.groom.orbit.goal.quest.dao.QuestRepository;
 import com.groom.orbit.goal.quest.dao.entity.Quest;
+import com.groom.orbit.infra.ai.app.VectorService;
+import com.groom.orbit.infra.ai.app.dto.CreateVectorDto;
+import com.groom.orbit.infra.ai.app.dto.UpdateVectorGoalDto;
 import com.groom.orbit.member.member.app.MemberQueryService;
 import com.groom.orbit.member.member.dao.jpa.entity.Member;
