+++ src/main/java/com/groom/orbit/goal/quest/app/QuestCommandService.java
@@ -5,9 +5,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.groom.orbit.ai.app.VectorService;
-import com.groom.orbit.ai.app.dto.CreateVectorDto;
-import com.groom.orbit.ai.app.dto.UpdateVectorQuestDto;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.goal.goal.app.MemberGoalService;
 import com.groom.orbit.goal.goal.dao.entity.MemberGoal;
@@ -16,6 +13,9 @@ import com.groom.orbit.goal.quest.app.dto.request.UpdateQuestRequestDto;
 import com.groom.orbit.goal.quest.app.dto.response.CreateQuestResponse;
 import com.groom.orbit.goal.quest.dao.QuestRepository;
 import com.groom.orbit.goal.quest.dao.entity.Quest;
+import com.groom.orbit.infra.ai.app.VectorService;
+import com.groom.orbit.infra.ai.app.dto.CreateVectorDto;
+import com.groom.orbit.infra.ai.app.dto.UpdateVectorQuestDto;
 
 import lombok.RequiredArgsConstructor;
