+++ src/main/java/com/groom/orbit/goal/quest/app/QuestRecommendService.java
@@ -2,10 +2,10 @@ package com.groom.orbit.goal.quest.app;
 
 import org.springframework.stereotype.Service;
 
-import com.groom.orbit.ai.app.AiService;
 import com.groom.orbit.goal.goal.app.MemberGoalService;
 import com.groom.orbit.goal.goal.dao.entity.MemberGoal;
 import com.groom.orbit.goal.quest.app.dto.response.RecommendQuestListResponseDto;
+import com.groom.orbit.infra.ai.app.AiService;
 
 import lombok.RequiredArgsConstructor;
