+++ src/main/java/com/groom/orbit/infra/ai/app/dto/CreateVectorDto.java
@@ -1,0 +1,13 @@
+package com.groom.orbit.infra.ai.app.dto;
+
+import java.util.List;
+
+import lombok.Builder;
+
+@Builder
+public record CreateVectorDto(
+    Long memberId,
+    String memberName,
+    List<String> interestJobs,
+    List<String> goals,
+    List<String> quests) {}
