+++ src/main/java/com/groom/orbit/infra/ai/app/dto/UpdateVectorGoalDto.java
@@ -1,0 +1,6 @@
+package com.groom.orbit.infra.ai.app.dto;
+
+import lombok.Builder;
+
+@Builder
+public record UpdateVectorGoalDto(Long memberId, String goal, String newGoal) {}
