+++ src/main/java/com/groom/orbit/infra/ai/app/dto/UpdateVectorQuestDto.java
@@ -1,0 +1,6 @@
+package com.groom.orbit.infra.ai.app.dto;
+
+import lombok.Builder;
+
+@Builder
+public record UpdateVectorQuestDto(Long memberId, String quest, String newQuest) {}
