+++ src/main/java/com/groom/orbit/infra/ai/app/util/PineconeConst.java
@@ -1,0 +1,13 @@
+package com.groom.orbit.infra.ai.app.util;
+
+public class PineconeConst {
+
+  private PineconeConst() {}
+
+  public static final String EMBEDDING_MODEL = "multilingual-e5-large";
+  public static final String INDEX_NAME = "goormthon";
+  public static final String INTEREST_JOB_NAMESPACE = "interest_job";
+  public static final String INPUT_TYPE_PARAMETER_KEY = "input_type";
+  public static final String QUERY_PARAMETER_VALUE = "query";
+  public static final String DEFAULT_MEMBER_NAME = "";
+}
