+++ src/main/java/com/groom/orbit/infra/fcm/FcmMessageDto.java
@@ -1,0 +1,8 @@
+package com.groom.orbit.infra.fcm;
+
+public record FcmMessageDto(Boolean validateOnly, Message message) {
+
+  public static FcmMessageDto of(Boolean validateOnly, Message message) {
+    return new FcmMessageDto(validateOnly, message);
+  }
+}
