+++ src/main/java/com/groom/orbit/infra/fcm/Message.java
@@ -1,0 +1,8 @@
+package com.groom.orbit.infra.fcm;
+
+public record Message(Notification notification, String token) {
+
+  public static Message of(Notification notification, String token) {
+    return new Message(notification, token);
+  }
+}
