+++ src/main/java/com/groom/orbit/infra/fcm/SaveFcmTokenRequestDto.java
@@ -1,0 +1,3 @@
+package com.groom.orbit.infra.fcm;
+
+public record SaveFcmTokenRequestDto(String fcmToken) {}
